+++ src/mmwave/model/mmwave-mac-pdu-header.h
@@ -38,7 +38,7 @@ class Tag;
 
 struct MacSubheader
 {
-	MacSubheader (uint8_t lcid, uint32_t size) :
+	MacSubheader (uint8_t lcid, uint16_t size) :
 		m_lcid (lcid), m_size (size)
 	{
 	}
